task_manager/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
PyJWT==2.8.0
drf-yasg==1.21.7